Added pattern test cases for few more detectors
